+++ js/utils.js
@@ -1,0 +1,17 @@
+const randomIntInRange = (left, right) => {
+  left = Math.ceil(left);
+  right = Math.floor(right);
+  return Math.floor(Math.random() * (right - left) + left);
+};
+
+const checkStrLength = (str, maxLength) => str.length <= maxLength;
+
+const createObject = (i, desc) => ({
+  id: i,
+  url: `photos/${i}.jpg`,
+  description: desc,
+  likes: randomIntInRange(15, 200),
+  comments: randomIntInRange(0, 200)
+});
+
+export {checkStrLength, randomIntInRange, createObject};
